ref(logs): Move Set Up Logs button from top bar to page filters
Move the SetupLogsButton from the TopBar actions slot to inline with
the page filters on the logs onboarding page. Extract it into its own
file to avoid a circular dependency between content.tsx and
logsOnboarding.tsx.
